requirements.txt
aiogram==3.20.0.post0
aiohttp==3.11.18
loguru==0.7.3
pydantic_settings==2.9.1
python-dotenv==1.1.0
Requests==2.32.3
